#43 Fix: 추천검색어 오류 수정 및 파일 분리
