+++ iTER/src/component/layout/TopSearch.tsx
@@ -2,12 +2,12 @@ import { styled } from '../../../stitches.config';
 import Back from '../../assets/icon/Back.svg?react';
 import XbtnCircle from '../../assets/icon/XbtnCircle.svg?react';
 import { useNavigate } from 'react-router-dom';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
+import RecommendItem from '../search/RecommendList';
 
 const TopSearch = ({
   back,
   onHandle,
-  onChange,
   searchText,
 }: {
   back?: () => void;
@@ -17,10 +17,7 @@ const TopSearch = ({
 }) => {
   const navigate = useNavigate();
   const [searchValue, setSearchValue] = useState<string>(searchText || '');
-
-  useEffect(() => {
-    setSearchValue(searchText || '');
-  }, [searchText]);
+  const [showSuggestions, setShowSuggestions] = useState(false);
 
   const handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>) => {
     if (event.key === 'Enter') {
@@ -29,6 +26,20 @@ const TopSearch = ({
     }
   };
 
+  const handleSelectSuggestions = (text: string) => {
+    setSearchValue(text);
+    setShowSuggestions(false);
+  };
+
+  const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const text = event.target.value;
+    setSearchValue(text);
+    if (text.trim() === '') {
+      setShowSuggestions(false);
+    } else {
+      setShowSuggestions(true);
+    }
+  };
   return (
     <>
       <Container>
@@ -40,18 +51,31 @@ const TopSearch = ({
             placeholder="검색어를 입력해 주세요"
             value={searchValue}
             onChange={(event) => {
-              setSearchValue(event.target.value);
-              onChange && onChange(event);
+              handleSearch(event);
             }}
             onKeyPress={handleKeyPress}
           />
           {searchValue && (
-            <Xbox onClick={() => setSearchValue('')}>
+            <Xbox
+              onClick={() => {
+                setSearchValue('');
+                setShowSuggestions(false);
+              }}
+            >
               <XbtnCircle />
             </Xbox>
           )}
         </InputBox>
       </Container>
+      {showSuggestions && searchValue?.length > 0 && (
+        <RecommendBox>
+          {recommend
+            .filter((text) => text.includes(searchValue))
+            .map((text) => (
+              <RecommendItem key={text} text={text} onSelect={handleSelectSuggestions} />
+            ))}
+        </RecommendBox>
+      )}
     </>
   );
 };
@@ -111,3 +135,14 @@ const Xbox = styled('div', {
   justifyContent: 'center',
   cursor: 'pointer',
 });
+
+const RecommendBox = styled('div', {
+  width: '390px',
+  height: '100vh',
+  marginTop: '60px',
+  backgroundColor: '$White',
+  position: 'absolute',
+  zIndex: 1,
+});
+
+const recommend = ['추천 검색어', '추천할까', '추천할게'];
